+++ CHANGELOG.md
@@ -109,6 +109,7 @@
 * Reconcile styling to latest Stripes styling. Refs [UILD-761].
 * Change behavior of "Assign" button between different lookups. Refs [UILD-745].
 * Add missing localization. Refs [UILD-784].
+* Add support for the Geographic Coverage field. Fixes [UILD-665].
 
 [UILD-552]:https://folio-org.atlassian.net/browse/UILD-552
 [UILD-544]:https://folio-org.atlassian.net/browse/UILD-544
@@ -217,6 +218,7 @@
 [UILD-761]:https://folio-org.atlassian.net/browse/UILD-761
 [UILD-745]:https://folio-org.atlassian.net/browse/UILD-745
 [UILD-784]:https://folio-org.atlassian.net/browse/UILD-784
+[UILD-665]:https://folio-org.atlassian.net/browse/UILD-665
 
 ## 1.0.5 (2025-04-30)
 * Browser Back button navigates to Edit without duplicated title when navigated from Duplicate screen. Fixes [UILD-554].